numerous updates
Added the necessary methods to check if a RummyBot has a winning hand (apologies for the StraightCheck method I really did try to make it legible) and fixed that one error in Deck for the billionth time, I'm not sure how it keeps coming back I think I might be using github slightly wrong. Also it's 3am right now so wheeeeeee! Time to go relearn graphics